# harvard-api-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/harvard/search":{"get":{"operationId":"HarvardController_searchArtworks","summary":"Buscar obras de arte","description":"\n      Permite buscar obras de arte en la colección del Harvard Art Museums.\n      \n      **Ejemplos de búsqueda:**\n      - `monet` - Busca obras de Claude Monet\n      - `paintings` - Busca pinturas\n      - `impressionist` - Busca arte impresionista\n      - `portrait` - Busca retratos\n      \n      **Características:**\n      - Búsqueda full-text en título, artista y descripción\n      - Resultados ordenados por relevancia\n      - Incluye imágenes cuando están disponibles\n      - Respuesta optimizada para AR Gallery\n    ","parameters":[{"name":"q","required":true,"in":"query","description":"Término de búsqueda (artista, obra, estilo, etc.)","schema":{"example":"monet","type":"string"}},{"name":"limit","required":false,"in":"query","description":"Número máximo de resultados a devolver","schema":{"example":"10","type":"string"}}],"responses":{"200":{"description":"Lista de obras de arte encontradas","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvardSearchResponseDTO"}}}},"400":{"description":"Parámetros de búsqueda inválidos"},"500":{"description":"Error interno del servidor"}},"tags":["harvard"]}},"/api/v1/harvard/artwork/{id}":{"get":{"operationId":"HarvardController_getArtworkById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["harvard"]}},"/api/v1/harvard/divisions":{"get":{"operationId":"HarvardController_getDivisions","parameters":[],"responses":{"200":{"description":""}},"tags":["harvard"]}},"/api/v1/harvard/classifications":{"get":{"operationId":"HarvardController_getClassifications","parameters":[],"responses":{"200":{"description":""}},"tags":["harvard"]}},"/api/v1/harvard/cultures":{"get":{"operationId":"HarvardController_getCultures","parameters":[],"responses":{"200":{"description":""}},"tags":["harvard"]}},"/api/v1/harvard/health":{"get":{"operationId":"HarvardController_healthCheck","summary":"Health check del servicio","description":"\n      Verifica el estado de salud del adaptador Harvard y su conectividad con la API externa.\n      \n      **Qué verifica:**\n      - Conectividad con Harvard Art Museums API\n      - Tiempo de respuesta del servicio\n      - Estado general del microservicio\n      \n      **Respuesta típica:**\n      - `healthy`: Servicio funcionando correctamente\n      - `unhealthy`: Problemas de conectividad o configuración\n      - `responseTime`: Tiempo en millisegundos de la verificación\n    ","parameters":[],"responses":{"200":{"description":"Estado de salud del servicio","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvardHealthResponseDTO"}}}},"500":{"description":"Error al verificar estado de salud"}},"tags":["harvard","health"]}}},"info":{"title":"Harvard Art Museums API","description":"Adaptador para búsqueda de obras de arte del Harvard Art Museums","version":"1.0.0","contact":{}},"tags":[{"name":"harvard","description":"Endpoints del Harvard Art Museums"},{"name":"health","description":"Endpoints de monitoreo y salud"}],"servers":[],"components":{"schemas":{"HarvardArtworkResponseDTO":{"type":"object","properties":{"objectId":{"type":"string","description":"ID único de la obra en la colección de Harvard","example":"331916"},"title":{"type":"string","description":"Título de la obra de arte","example":"Water Lilies"},"artistName":{"type":"string","description":"Nombre del artista o creador","example":"Claude Monet"},"imageUrl":{"type":"string","description":"URL de la imagen de la obra (si está disponible)","example":"https://nrs.harvard.edu/urn-3:HUAM:331916"},"department":{"type":"string","description":"Departamento o división del museo","example":"European and American Art"},"culture":{"type":"string","description":"Cultura o origen de la obra","example":"French"},"period":{"type":"string","description":"Período histórico de la obra","example":"Early 20th century"},"century":{"type":"string","description":"Siglo de creación","example":"20th century"},"dated":{"type":"string","description":"Fecha de creación de la obra","example":"1919"},"medium":{"type":"string","description":"Técnica y materiales utilizados","example":"Oil on canvas"},"dimensions":{"type":"string","description":"Dimensiones de la obra","example":"100.3 x 81.3 cm"},"isPublicDomain":{"type":"boolean","description":"Si la obra está en dominio público","example":true},"objectURL":{"type":"string","description":"URL con información detallada de la obra","example":"https://harvardartmuseums.org/collections/object/331916"},"tags":{"description":"Etiquetas asociadas a la obra","example":["impressionism","landscape","water lilies"],"type":"array","items":{"type":"string"}},"description":{"type":"string","description":"Descripción detallada de la obra","example":"One of Monet's famous water lily paintings from his garden in Giverny"},"accessionNumber":{"type":"string","description":"Número de registro de la obra","example":"1979.277"},"classification":{"type":"string","description":"Clasificación artística","example":"Paintings"},"technique":{"type":"string","description":"Técnica específica utilizada","example":"Oil painting"},"provenance":{"type":"string","description":"Procedencia y historial de la obra","example":"Donated by the artist's family in 1979"},"copyright":{"type":"string","description":"Información de derechos de autor","example":"Public Domain"},"museum":{"type":"string","description":"Nombre del museo","example":"Harvard Art Museums","default":"Harvard Art Museums"},"hasImage":{"type":"boolean","description":"Indica si la obra tiene imagen disponible","example":true}},"required":["objectId","title","artistName","imageUrl","department","museum","hasImage"]},"HarvardSearchResponseDTO":{"type":"object","properties":{"query":{"type":"string","description":"Término de búsqueda utilizado","example":"monet"},"total":{"type":"number","description":"Número total de obras encontradas","example":86},"objectIDs":{"description":"Lista de IDs de las obras encontradas","example":["331916","227751","228125"],"type":"array","items":{"type":"string"}},"artworks":{"description":"Lista de obras de arte con información detallada","type":"array","items":{"$ref":"#/components/schemas/HarvardArtworkResponseDTO"}},"executionTimeMs":{"type":"number","description":"Tiempo de ejecución de la búsqueda en millisegundos","example":245},"limit":{"type":"number","description":"Número de resultados solicitados","example":10},"filters":{"type":"object","description":"Filtros aplicados en la búsqueda","example":{"classification":"Paintings","hasImage":true}}},"required":["query","total","objectIDs","artworks","executionTimeMs"]},"HarvardHealthResponseDTO":{"type":"object","properties":{"status":{"type":"string","description":"Estado de salud del servicio","example":"healthy","enum":["healthy","unhealthy"]},"timestamp":{"type":"string","description":"Timestamp del health check en formato ISO","example":"2025-10-22T07:30:45.123Z"},"service":{"type":"string","description":"Nombre del servicio","example":"harvard-adapter","default":"harvard-adapter"},"version":{"type":"string","description":"Versión del servicio","example":"1.0.0","default":"1.0.0"},"harvardApiConnection":{"type":"boolean","description":"Estado de la conexión con Harvard Art Museums API","example":true},"responseTimeMs":{"type":"number","description":"Tiempo de respuesta del health check en millisegundos","example":245}},"required":["status","timestamp","service","version","harvardApiConnection"]}}}}